extract Executor concept which can be provided separately from basic appbase functionality.
